Move ShortUrl model to its own file Rename app.rb to shortnatra.rb
